requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.31
flask-sqlalchemy==3.0.1
python-dotenv==0.21.0